+++ src/index.ts
@@ -14,4 +14,4 @@ export type {
 /**
  * パッケージの公開版数。
  */
-export const version = '0.5.3';
+export const version = '0.5.4';